+++ RustyBags/Utilities/Keys.cs
@@ -36,6 +36,7 @@ public static class Keys
     public static readonly string OreWeight = new Key("$se_oreweight", "Ore Weight").key;
     public static readonly string Bag = new Keys.Key("$label_bag", "Bag").key;
     public static readonly string SECharm = new Key("$se_charm", "No bag speed penalty").key;
+    public static readonly string IfEquippedToBag = new Key("$se_charm_equipped", "If equipped to bag").key;
     
     public const string Blunt = "$inventory_blunt";
     public const string Slash = "$inventory_slash";